Company owner manages users for his company

+++ app/Http/Controllers/CompanyUserController.php
@@ -12,6 +12,8 @@ class CompanyUserController extends Controller
 {
     public function index(Company $company)
     {
+        $this->authorize('viewAny', $company);
+
         $users = $company->users()->get();
 
         return view('companies.users.index', compact('company', 'users'));
@@ -19,16 +21,20 @@ class CompanyUserController extends Controller
 
     public function create(Company $company)
     {
+        $this->authorize('create', $company);
+
         return view('companies.users.create', compact('company'));
     }
 
     public function store(StoreUserRequest $request, Company $company)
     {
+        $this->authorize('create', $company);
+
         $company->users()->create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
             'password' => bcrypt($request->input('password')),
-            'role_id' => Role::CUSTOMER,
+            'role_id' => Role::CUSTOMER->value,
         ]);
 
         return to_route('companies.users.index', $company);
@@ -36,11 +42,15 @@ class CompanyUserController extends Controller
 
     public function edit(Company $company, User $user)
     {
+        $this->authorize('update', $company);
+
         return view('companies.users.edit', compact('company', 'user'));
     }
 
     public function update(UpdateUserRequest $request, Company $company, User $user)
     {
+        $this->authorize('update', $company);
+
         $user->update($request->validated());
 
         return to_route('companies.users.index', $company);
@@ -48,6 +58,8 @@ class CompanyUserController extends Controller
 
     public function destroy(Company $company, User $user)
     {
+        $this->authorize('delete', $company);
+
         $user->delete();
 
         return to_route('companies.users.index', $company);
